requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine Learning / Deep Learning
# Note: TensorFlow not yet available for Python 3.14, use Python 3.11 or 3.12 for full ML features
# tensorflow>=2.13.0
# keras>=2.13.0
# torch>=2.0.0
scikit-learn>=1.3.0

# Power System Analysis
pandapower>=2.13.0
pypower>=5.1.0

# Optimization
cvxpy>=1.3.0
pyomo>=6.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Processing
openpyxl>=3.1.0
h5py>=3.8.0

# MATLAB Integration
matlabengine>=9.14.0  # Optional - requires MATLAB installation

# Cybersecurity & Network
scapy>=2.5.0
pycryptodome>=3.18.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.2.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# PDF Processing
pypdf>=3.0.0
